requirements/base.txt
Django==1.8.5
#bpython==0.13
django-braces==1.8.1
django-model-utils==2.3.1
logutils==0.3.3
#psycopg2==2.6.1
static==1.1.1
dj-database-url==0.3.0
dj-static==0.0.6
#django-debug-toolbar==1.2.1
django-autoslug==1.9.3
openpyxl==2.3.0
django-datetime-widget==0.9.3
djangorestframework==3.3.0
markdown==2.6.3
django-filter==0.11.0
django-bootstrap3==6.2.2
jsonfield==1.0.3
celery==3.1.19
django-celery==3.1.17